Wait for the RPC Server to start before testing
When creating the RPC Server thread, a dummy event is added to the
event loop. When the event loop has fully started, the dummy event
unregisters itself and allows setup_env() to return.

This fixes a race condition on my system, where test_connect_failed
would fail so fast that the server ioloop didn't start by the time
the test teardown would request the ioloop to stop.